Add files via upload
The Html pages will show you the code. The HTML screenshots will also show the code as a picture. Also since some of my pages are a bit long, I will screenshot parts of the code for you to see. Some of the screenshots are the pages itself and some of them are code screenshots from Brackets.